Update template to support paging.
We want to be able to page through results so we need a call that accepts a URI to allow us to pass through the URLs from the link headers canvas returns. There's no way to have just one method for each model used so instead I just overloaded all calls that return a list to have a version that can be used for paging.

I also added a line break in on the class declaration.